Create project store
- Create store for projects
- Move project related logic to store
